backend/ocr/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
pdf2image
python-dotenv 